refactor: callLlama component
we need test!!!
