+++ nuxt-llama-gpt/app/pages/chat/[id].vue
@@ -17,7 +17,7 @@
   <v-bottom-navigation :name="'navigation' + route.params.id" height="94" elevation="0" bg-color="#ffffff00" >
     <v-row justify="center">
       <v-col cols="8">
-        <CallLlama @sendMessage="callLlama"/>
+        <CallLlama v-model="userInput" @sendMessage="callLlama"/>
       </v-col>
     </v-row>
   </v-bottom-navigation>
@@ -26,7 +26,14 @@
 </template>
 
 <script setup lang="ts">
+
+/**
+ * 1. 클릭해서 넘어올때
+ * 2. 인덱스에서 세션만들고 넘어올때 
+ */
+
 import type { DefineComponent } from 'vue';
+import { CustomHttpCode } from '~/common/custom-http-code';
 import { useHelper } from '~/stores/useHelper';
 import type { Chat } from '~~/server/models/ChatHistory';
 const ChatClient = markRaw(defineAsyncComponent(() =>
@@ -44,13 +51,32 @@ console.log(data.value)
 
 const helper = useHelper()
 const { indexSay } = storeToRefs(helper)
-
 const socket = useSocket()
 const { $socket } = storeToRefs(socket)
 const goTo = useGoTo()
 const chatAreaEl = ref()
 const { height } = useElementSize(chatAreaEl)
+
+const userInput = ref()
 const word = ref()
+
+const { execute: textExecute } = useLazyFetch(() => `/api/llama/create-text/${route.params.id}`,{
+  method: 'POST',
+  watch: false,
+  immediate: false,
+  body:{
+    message: userInput  // input에 있는 text미리 없애기 
+  },
+  onResponseError: ({ request, response, options }) => {
+    const { status } = response
+    const { openModal } = usePageAuth()
+    
+    if(status == CustomHttpCode.LoginSessionInvailed) {
+      openModal()
+    }
+  }
+})
+
 watchEffect(() => {
   if($socket.value){
     $socket.value.on('chat', mess => {
@@ -63,15 +89,20 @@ watchEffect(() => {
   }
 })
 onMounted(() => {
-  // index에서 넘어올때 작동
+  // index에서 넘어올때 작동 => 필요할지도...
   if(indexSay.value){
-    // console.log('너뭐해 ', indexSay.value)
-    callLlama(indexSay.value)
+    userInput.value = indexSay.value
+    callLlama()
     indexSay.value = ''
   }
   // 세션을 누르고 들어올때 작동 
+  restoreChatHistory()
+})
+
+const restoreChatHistory = () => {
   if(!data.value) return 
   
+  // 기존 chat세션에 남아있는 히스토리 가져와서 배치해줌 
   data.value.forEach(history => {
     history.messages.forEach(message => {
       if (message.type === 'user') {
@@ -91,18 +122,16 @@ onMounted(() => {
     // history.messages[1] // user
     // history.messages[2] // say llama
   })
-
-
-})
+}
 
 const contentList = ref<ChatArea[]>([])
 
 
 // 클라입력 직후에 바로 ai 답변대기 표시를 하고 그 다음에 socket으로 답변을 받는다.
-const callLlama = async(say: string) => {
+const callLlama = async() => {
   contentList.value.push({
     component: ChatClient,
-    saying: say,
+    saying: userInput.value,
   })
 
   word.value = ''
@@ -115,6 +144,9 @@ const callLlama = async(say: string) => {
 
   await delay(1000)
   goTo(height.value)
+
+  // goto 후에 했던거 같은데 
+  await textExecute()
 }
